updated eventing upgrade-job to ignore 1.23.x
