+++ components/eventing-controller/upgrade-job/clients/eventingbackend/eventingbackend.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"encoding/json"
 
-	"k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
-	apixv1beta1client "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/typed/apiextensions/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -16,13 +14,12 @@ import (
 
 // Client struct for EventingBackend client
 type Client struct {
-	client              dynamic.Interface
-	apiExtensionsClient *apixv1beta1client.ApiextensionsV1beta1Client
+	client dynamic.Interface
 }
 
 // NewClient creates and returns new client for EventingBackend
-func NewClient(client dynamic.Interface, apiExtensionsClient *apixv1beta1client.ApiextensionsV1beta1Client) Client {
-	return Client{client, apiExtensionsClient}
+func NewClient(client dynamic.Interface) Client {
+	return Client{client}
 }
 
 // Get returns the EventingBackend for specified name and namespace.
@@ -36,17 +33,6 @@ func (c Client) Get(namespace string, name string) (*eventingv1alpha1.EventingBa
 		return nil, err
 	}
 	return toEventingBackend(ebUnstructured)
-}
-
-// GetCRD returns the EventingBackend CRD
-// or returns an error if the EventingBackend CRD was not found or other issues
-func (c Client) GetCRD() (*v1beta1.CustomResourceDefinition, error) {
-	crd, err := c.apiExtensionsClient.CustomResourceDefinitions().Get(context.Background(), "eventingbackends.eventing.kyma-project.io", metav1.GetOptions{})
-
-	if err != nil {
-		return nil, err
-	}
-	return crd, nil
 }
 
 // GroupVersionResource return the GVR for EventingBackend k8s resource
